test: move file descriptor leak checking to common
